feat: Expose binary as CLI
