+++ crates/stream_pulse/Cargo.toml
@@ -6,8 +6,8 @@ edition.workspace = true
 categories.workspace = true
 
 [[bin]]
-name = "stream-pulse-cron"
-path = "./src/bin/stream-pulse-cron.rs"
+name = "stream-pulse"
+path = "./src/bin/stream-pulse.rs"
 
 [lib]
 name = "stream_pulse"
@@ -24,6 +24,7 @@ apalis = { version = "1.0.0-rc.3", features = [
 apalis-cron = "1.0.0-rc.3"
 chrono = { workspace = true }
 chrono-tz = "0.10.0"
+clap = { version = "4.5.40", features = ["derive", "env"] }
 cron = "0.15.0"
 dotenvy = "0.15.7"
 futures = "0.3.30"
